+++ src/appReducer.ts
@@ -9,7 +9,6 @@ import { combineReducers } from 'redux';
 
 import chartSlice from './components/EventChart/Chart/chartSlice';
 import modemReducer from './features/modem/modemSlice';
-import powerEstimationReducer from './features/powerEstimation/powerEstimationSlice';
 import traceReducer from './features/tracing/traceSlice';
 import dashboardReducer from './features/tracingEvents/dashboardSlice';
 import wiresharkReducer from './features/wireshark/wiresharkSlice';
@@ -21,7 +20,6 @@ export type RootState = NrfConnectState<AppState>;
 const appReducer = combineReducers({
     modem: modemReducer,
     trace: traceReducer,
-    powerEstimation: powerEstimationReducer,
     wireshark: wiresharkReducer,
     dashboard: dashboardReducer,
     chart: chartSlice,