+++ src/components/SidePanel/SidePanel.tsx
@@ -12,18 +12,11 @@ import AdvancedOptions from './AdvancedOptions';
 import ConnectionStatus from './ConnectionStatus';
 import EventGraphOptions from './EventGraphOptions';
 import Instructions from './Instructions';
-import PowerEstimationParams from './PowerEstimationParams';
 import TraceCollector from './Tracing/TraceCollector';
 import TraceFileInformation from './Tracing/TraceFileInformation';
 
 import './sidepanel.scss';
 import './Tracing/tracing.scss';
-
-export const PowerEstimationSidePanel = () => (
-    <SidePanel>
-        <PowerEstimationParams />
-    </SidePanel>
-);
 
 export const TraceCollectorSidePanel = () => (
     <SidePanel className="side-panel">