+++ src/features/tracing/nrfml.test.ts
@@ -9,7 +9,6 @@ import { testUtils } from 'pc-nrfconnect-shared/test';
 
 import appReducer from '../../appReducer';
 import { getMockStore, mockedDataDir } from '../../utils/testUtils';
-import { resetParams as resetPowerEstimationParams } from '../powerEstimation/powerEstimationSlice';
 import { resetDashboardState } from '../tracingEvents/dashboardSlice';
 import nrfml from './__mocks__/@nordicsemiconductor/nrf-monitor-lib-js';
 import { convertTraceFile, startTrace } from './nrfml';
@@ -87,7 +86,6 @@ describe('nrfml', () => {
         it('should start tracing to pcap', () => {
             store.dispatch(startTrace(['pcap']));
             expect(store.getActions()).toEqual([
-                { type: resetPowerEstimationParams.type, payload: undefined },
                 { type: resetDashboardState.type, payload: undefined },
                 { type: setTraceSourceFilePath.type, payload: null },
                 { type: setTraceDataReceived.type, payload: false },
@@ -113,7 +111,6 @@ describe('nrfml', () => {
         it('should start tracing to raw binary', () => {
             store.dispatch(startTrace(['raw']));
             expect(store.getActions()).toEqual([
-                { type: resetPowerEstimationParams.type, payload: undefined },
                 { type: resetDashboardState.type, payload: undefined },
                 { type: setTraceSourceFilePath.type, payload: null },
                 { type: setTraceDataReceived.type, payload: false },
@@ -138,7 +135,6 @@ describe('nrfml', () => {
         it('does not create a progress config for live traces', () => {
             store.dispatch(startTrace(['raw', 'live']));
             expect(store.getActions()).toEqual([
-                { type: resetPowerEstimationParams.type, payload: undefined },
                 { type: resetDashboardState.type, payload: undefined },
                 { type: setTraceSourceFilePath.type, payload: null },
                 { type: setTraceDataReceived.type, payload: false },
@@ -164,7 +160,6 @@ describe('nrfml', () => {
         it('does not create a progress config for live traces', () => {
             store.dispatch(startTrace(['live']));
             expect(store.getActions()).toEqual([
-                { type: resetPowerEstimationParams.type, payload: undefined },
                 { type: resetDashboardState.type, payload: undefined },
                 { type: setTraceSourceFilePath.type, payload: null },
                 { type: setTraceDataReceived.type, payload: false },